requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
install==1.3.5
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2024.1
urllib3==2.2.1
